# src/main/webui/openapi.json
{
  "openapi": "3.1.0",
  "components": {
    "schemas": {
...
      "ReportData": {
        "description": "A single report request",
        "type": "object",
        "required": [
          "reportRequestId",
          "report"
        ],
        "properties": {
          "reportRequestId": {
            "description": "Report identification",
            "type": "object",
            "required": [
              "id",
              "reportId"
            ],
            "properties": {
              "id": {
                "type": "string",
                "description": "ID assigned for database identification"
              },
              "reportId": {
                "type": "string",
                "description": "ID assigned on scan request submission"
              }
            }
          },
          "report": {
            "description": "Report payload",
            "type": "object"
          }
        }
      },
...
      "ReportRequestId": {
        "description": "Report identification",
        "type": "object",
        "required": [
          "id",
          "reportId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "ID assigned for database identification"
          },
          "reportId": {
            "type": "string",
            "description": "ID assigned on scan request submission"
          }
        }
      },
      "ReportWithStatus": {
        "description": "Report data with calculated analysis status",
        "type": "object",
        "required": [
          "report",
          "status"
        ],
        "properties": {
          "report": {
            "description": "Full report data from MongoDB",
            "type": "object"
          },
          "status": {
            "type": "string",
            "description": "Calculated analysis status of the report (completed, queued, sent, expired, failed, pending, unknown)"
          }
        }
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "description": "Please enter your JWT Token without Bearer",
        "scheme": "bearer",
        "bearerFormat": "jwt"
      },
      "SecurityScheme": {
        "type": "openIdConnect",
        "description": "Authentication"
      }
    }
  },
  "paths": {
...
    "/api/v1/reports": {
      "delete": {
        "summary": "Delete multiple analysis reports",
        "description": "Deletes multiple analysis reports by IDs or using filter parameters",
        "parameters": [
          {
            "description": "List of report IDs to delete (24-character hexadecimal MongoDB ObjectId format)",
            "name": "reportIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Reports deletion request accepted"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      },
      "post": {
        "summary": "Receive analysis report",
        "description": "Receives a completed analysis report from Morpheus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Report received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRequestId"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      },
      "get": {
        "summary": "List analysis reports",
        "description": "Retrieves a paginated list of analysis reports with optional filtering and sorting",
        "parameters": [
          {
            "description": "Filter by ExploitIQ status. Valid values: TRUE, FALSE, UNKNOWN",
            "name": "exploitIqStatus",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by image name",
            "name": "imageName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by image tag",
            "name": "imageTag",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Page number (0-based)",
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "0"
            }
          },
          {
            "description": "Number of items per page",
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "100"
            }
          },
          {
            "description": "Filter by SBOM report ID (metadata.product_id)",
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by report ID (input.scan.id)",
            "name": "reportId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort criteria in format 'field:direction'",
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": "submittedAt:DESC"
            }
          },
          {
            "description": "Filter by status. Valid values: completed, sent, failed, queued, expired, pending",
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by vulnerability ID (CVE ID)",
            "name": "vulnId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reports retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Report"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      }
    },
    "/api/v1/reports/new": {
      "post": {
        "summary": "Create new analysis request",
        "description": "Creates a new analysis report request, processes it and optionally submits it to ExploitIQ for analysis",
        "parameters": [
          {
            "description": "Whether to submit to ExploitIQ for analysis",
            "name": "submit",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "requestBody": {
          "description": "Analysis report request data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Analysis request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportData"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "429": {
            "description": "Request queue exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      }
    },
    "/api/v1/reports/product": {
      "delete": {
        "summary": "Delete product by IDs",
        "description": "Deletes all component analysis reports and product metadata associated with specified product IDs",
        "parameters": [
          {
            "description": "List of product IDs to delete",
            "required": true,
            "name": "productIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Product deletion request accepted"
          },
          "400": {
            "description": "Invalid request - no product IDs provided"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      },
      "get": {
        "summary": "List all product data",
        "description": "Retrieves paginated, sortable, and filterable product data for all products",
        "parameters": [
          {
            "name": "cveId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "0"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": "100"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "DESC"
            }
          },
          {
            "name": "sortField",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "submittedAt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductSummary"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      }
    },
...
    "/api/v1/reports/{id}": {
      "delete": {
        "summary": "Delete analysis report",
        "description": "Deletes a specific analysis report by ID",
        "parameters": [
          {
            "description": "Report ID to delete (24-character hexadecimal MongoDB ObjectId format)",
            "required": true,
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Report deletion request accepted"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      },
      "get": {
        "summary": "Get analysis report",
        "description": "Retrieves a specific analysis report by ID with calculated analysis status",
        "parameters": [
          {
            "description": "Report ID to get (24-character hexadecimal MongoDB ObjectId format)",
            "required": true,
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Report retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportWithStatus"
                }
              }
            }
          },
          "404": {
            "description": "Report not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Report Endpoint"
        ]
      }
    },
...
  },
...
  "info": {
    "title": "exploit-iq-client API",
    "version": "1.0.0-SNAPSHOT"
  },
  "servers": []
}